feat(GasFees): update all places where gas fees are required in the exchange UI
