+++ src/pages/Options/Market/TradeCard/components/BidNetworkFees/BidNetworkFees.tsx
@@ -4,13 +4,12 @@ import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import Tooltip from '@material-ui/core/Tooltip';
 
-import { TextButton, FlexDivRow, GridDivRow, FlexDivCentered, FlexDiv } from 'shared/commonStyles';
+import { FlexDivRow, GridDivRow, FlexDivCentered, FlexDiv } from 'shared/commonStyles';
 import { DataSmall } from 'components/Typography';
 
 import { formatCurrencyWithSign, formatPercentage } from 'utils/formatters';
 import { getTransactionPrice } from 'utils/networkUtils';
 
-import { toggleGweiPopup } from 'ducks/ui';
 import { getEthRate, getRatesExchangeRates } from 'ducks/rates';
 import { getGasInfo } from 'ducks/transaction';
 import { RootState } from 'ducks/types';
@@ -21,17 +20,15 @@ import { USD_SIGN } from 'constants/currency';
 import { formDataCSS } from 'components/Typography/Form';
 import { ReactComponent as QuestionMark } from 'assets/images/question-mark.svg';
 
+import SelectGasMenu from 'pages/shared/components/SelectGasMenu';
+
 const mapStateToProps = (state: RootState) => ({
 	exchangeRates: getRatesExchangeRates(state),
 	gasInfo: getGasInfo(state),
 	ethRate: getEthRate(state),
 });
 
-const mapDispatchToProps = {
-	toggleGweiPopup,
-};
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
+const connector = connect(mapStateToProps);
 
 type PropsFromRedux = ConnectedProps<typeof connector>;
 
@@ -45,7 +42,6 @@ type NetworkFeesProps = PropsFromRedux & {
 
 const NetworkFees: FC<NetworkFeesProps> = ({
 	className,
-	toggleGweiPopup,
 	gasInfo,
 	ethRate,
 	gasLimit,
@@ -100,10 +96,7 @@ const NetworkFees: FC<NetworkFeesProps> = ({
 			</FlexDivRow>
 			<FlexDivRow>
 				<div>{t('common.gas-price-gwei')}</div>
-				<div>
-					{gasPrice || 0}
-					<ButtonEdit onClick={() => toggleGweiPopup(true)}>{t('common.actions.edit')}</ButtonEdit>
-				</div>
+				<SelectGasMenu gasPrice={gasPrice} />
 			</FlexDivRow>
 		</Container>
 	);
@@ -115,13 +108,6 @@ export const Container = styled(GridDivRow)`
 	text-transform: uppercase;
 	grid-gap: 8px;
 	padding-bottom: 16px;
-`;
-
-export const ButtonEdit = styled(TextButton)`
-	${formDataCSS};
-	margin-left: 10px;
-	color: ${(props) => props.theme.colors.buttonHover};
-	text-transform: uppercase;
 `;
 
 const QuestionMarkIcon = styled.div`
