+++ src/pages/Options/components/NetworkFees/NetworkFees.tsx
@@ -3,12 +3,11 @@ import { ConnectedProps, connect } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import { TextButton, FlexDivRow, GridDivRow } from 'shared/commonStyles';
+import { FlexDivRow, GridDivRow } from 'shared/commonStyles';
 
 import { formatCurrencyWithSign } from 'utils/formatters';
 import { getTransactionPrice } from 'utils/networkUtils';
 
-import { toggleGweiPopup } from 'ducks/ui';
 import { getEthRate, getRatesExchangeRates } from 'ducks/rates';
 import { getGasInfo } from 'ducks/transaction';
 import { RootState } from 'ducks/types';
@@ -17,17 +16,15 @@ import { USD_SIGN } from 'constants/currency';
 
 import { formDataCSS } from 'components/Typography/Form';
 
+import SelectGasMenu from 'pages/shared/components/SelectGasMenu';
+
 const mapStateToProps = (state: RootState) => ({
 	exchangeRates: getRatesExchangeRates(state),
 	gasInfo: getGasInfo(state),
 	ethRate: getEthRate(state),
 });
 
-const mapDispatchToProps = {
-	toggleGweiPopup,
-};
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
+const connector = connect(mapStateToProps);
 
 type PropsFromRedux = ConnectedProps<typeof connector>;
 
@@ -36,13 +33,7 @@ type NetworkFeesProps = PropsFromRedux & {
 	gasLimit: number | null;
 };
 
-const NetworkFees: FC<NetworkFeesProps> = ({
-	className,
-	toggleGweiPopup,
-	gasInfo,
-	ethRate,
-	gasLimit,
-}) => {
+const NetworkFees: FC<NetworkFeesProps> = ({ className, gasInfo, ethRate, gasLimit }) => {
 	const { t } = useTranslation();
 
 	const { gasPrice } = gasInfo;
@@ -57,10 +48,7 @@ const NetworkFees: FC<NetworkFeesProps> = ({
 			</FlexDivRow>
 			<FlexDivRow>
 				<div>{t('common.gas-price-gwei')}</div>
-				<div>
-					{gasPrice || 0}
-					<ButtonEdit onClick={() => toggleGweiPopup(true)}>{t('common.actions.edit')}</ButtonEdit>
-				</div>
+				<SelectGasMenu gasPrice={gasPrice} />
 			</FlexDivRow>
 		</Container>
 	);
@@ -72,13 +60,6 @@ export const Container = styled(GridDivRow)`
 	text-transform: uppercase;
 	grid-gap: 8px;
 	padding-bottom: 16px;
-`;
-
-export const ButtonEdit = styled(TextButton)`
-	${formDataCSS};
-	margin-left: 10px;
-	color: ${(props) => props.theme.colors.icons};
-	text-transform: uppercase;
 `;
 
 export default connector(NetworkFees);
