+++ src/pages/Trade/components/CreateOrderCard/NetworkInfo.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
@@ -7,13 +7,10 @@ import Tooltip from '@material-ui/core/Tooltip';
 import { formatCurrency } from 'utils/formatters';
 import { getTransactionPrice } from 'utils/networkUtils';
 
-import { LinkTextSmall } from 'shared/commonStyles';
 import { FlexDivRow } from 'shared/commonStyles';
-import { media } from 'shared/media';
 
-import GasMenu from 'pages/shared/components/GasMenu';
+import SelectGasMenu from 'pages/shared/components/SelectGasMenu';
 
-import DropdownPanel from 'components/DropdownPanel';
 import { DataSmall } from 'components/Typography';
 
 import { ReactComponent as QuestionMark } from 'assets/images/question-mark.svg';
@@ -28,14 +25,6 @@ export const TransactionInfo = ({
 	exchangeFeeRate = 0,
 }) => {
 	const { t } = useTranslation();
-
-	const [gasDropdownIsOpen, setGasDropdownIsOpen] = useState(false);
-	const setDropdownIsOpen = (isOpen) => {
-		if (!isOpen && !gasDropdownIsOpen) {
-			return;
-		}
-		setGasDropdownIsOpen(isOpen);
-	};
 
 	const usdValue = amount * usdRate;
 	const exchangeFee = ((amount * exchangeFeeRate) / 100) * usdRate;
@@ -76,20 +65,7 @@ export const TransactionInfo = ({
 			<NetworkDataRow>
 				<NetworkData>{t('common.gas-price-gwei')}</NetworkData>
 				<NetworkData>
-					<StyledDropdownPanel
-						height="auto"
-						isOpen={gasDropdownIsOpen}
-						handleClose={() => setDropdownIsOpen(false)}
-						width="150px"
-						onHeaderClick={() => setDropdownIsOpen(!gasDropdownIsOpen)}
-						header={
-							<GasEditFields>
-								<>{formatCurrency(gasPrice) || 0}</>
-								<LinkTextSmall>{t('common.actions.edit')}</LinkTextSmall>
-							</GasEditFields>
-						}
-						body={<GasMenu setDropdownIsOpen={setDropdownIsOpen} />}
-					/>
+					<SelectGasMenu gasPrice={gasPrice} />
 				</NetworkData>
 			</NetworkDataRow>
 		</Container>
@@ -101,24 +77,6 @@ TransactionInfo.propTypes = {
 	gasLimit: PropTypes.number,
 	ethRate: PropTypes.number,
 };
-
-const StyledDropdownPanel = styled(DropdownPanel)`
-	${media.medium`
-		width: auto;
-	`}
-	.body {
-		border-width: 1px 1px 1px 1px;
-		margin-top: 20px;
-	}
-`;
-
-const GasEditFields = styled.div`
-	display: flex;
-	width: 75px;
-	float: right;
-	justify-content: space-between;
-	cursor: pointer;
-`;
 
 const TooltipContent = styled.div`
 	width: 200px;
