--- src/pages/shared/components/GasMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './GasMenu';
