+++ src/pages/shared/components/SelectGasMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectGasMenu';
